fixed regression from the previous commit - when the user leaves the create-contact route, the record should be deleted from the store if it's not saved on the server

+++ app/components/address-book/create/component.js
@@ -55,7 +55,7 @@ export default Ember.Component.extend(countries, notifier, {
   * @description Deletes the model if it's not created or already deleted
   */
   deleteModel() {
-    if (this.get('model').get('id')) {
+    if (!this.get('model').get('isDeleted') && !this.get('model').get('id')) {
       this.get('model').deleteRecord();
     }
   },
